Move notification settings JSON parsing to CloudConnection.
Keeps URL definition and parsing code together.
